-- Major refactoring to System and Custom method loading.
The goal of this refactoring is to remove large switch cases for running
methods in the console. Also, the console methods are now located in the Leikr directory for extending by end users. I may abstract these with an API down the road to reduce the amount of infrastructure code end users will deal with.